# v1-simple/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Comments API
  version: 1.0.0
  description: API for managing comments (MVP). Supports CRUD operations, reading by post, and like management.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Comments
    description: Comment operations

paths:
  /comments:
    post:
      tags:
        - Comments
      summary: Create a comment
      operationId: createComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'

  /comments/{commentId}:
    get:
      tags:
        - Comments
      summary: Get comment by ID
      operationId: getCommentById
      parameters:
        - $ref: '#/components/parameters/commentId'
      responses:
        '200':
          description: Comment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Comments
      summary: Update comment text
      operationId: updateCommentText
      parameters:
        - $ref: '#/components/parameters/commentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentTextUpdate'
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Comments
      summary: Delete a comment
      operationId: deleteComment
      parameters:
        - $ref: '#/components/parameters/commentId'
      responses:
        '204':
          description: Comment deleted successfully (no content)
        '404':
          $ref: '#/components/responses/NotFound'

  /comments/{commentId}/like:
    post:
      tags:
        - Comments
      summary: Increment like count by 1
      operationId: incrementLikes
      parameters:
        - $ref: '#/components/parameters/commentId'
      responses:
        '200':
          description: Like count incremented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Comments
      summary: Decrement like count by 1
      operationId: decrementLikes
      parameters:
        - $ref: '#/components/parameters/commentId'
      responses:
        '200':
          description: Like count decremented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Cannot decrement likes below zero
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

  /posts/{postId}/comments:
    get:
      tags:
        - Comments
      summary: Get comments for a post
      operationId: getCommentsByPost
      parameters:
        - name: postId
          in: path
          required: true
          description: Post identifier
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          description: Maximum number of comments to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          description: Pagination offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'

components:
  parameters:
    commentId:
      name: commentId
      in: path
      required: true
      description: Comment identifier
      schema:
        type: integer
        format: int64

  schemas:
    Comment:
      type: object
      required:
        - id
        - author
        - postId
        - text
        - likes
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique comment identifier
        author:
          type: integer
          format: int64
          description: Author identifier
        postId:
          type: integer
          format: int64
          description: Post identifier this comment belongs to
        text:
          type: string
          minLength: 1
          maxLength: 1024
          description: Comment text
        likes:
          type: integer
          minimum: 0
          description: Like count
        createdAt:
          type: string
          format: date-time
          description: Creation time in ISO 8601 format

    CommentCreate:
      type: object
      required:
        - author
        - postId
        - text
      properties:
        author:
          type: integer
          format: int64
          description: Author identifier
        postId:
          type: integer
          format: int64
          description: Post identifier
        text:
          type: string
          minLength: 1
          maxLength: 1024
          description: Comment text
        likes:
          type: integer
          minimum: 0
          default: 0
          description: Initial like count (default 0)

    CommentTextUpdate:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1024
          description: New comment text

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error description
          example: "Likes count cannot be negative"

  responses:
    BadRequest:
      description: Invalid request (e.g., invalid data)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'